pratybos

+++ keyboard.h
@@ -10,7 +10,7 @@ private:
 	std::vector<char> keysBuffer;
 public:
 	uint8_t getKey();
-	void getBytes(uint8_t* dest, uint32_t rnum);
+	void getBytes(uint32_t* dest, uint32_t rnum);
 };
 
 #endif
