+++ monitor.h
@@ -7,10 +7,10 @@
 
 class Monitor {
 public:
-	void display(const uint8_t* src, uint32_t bytes);
-	void display(const uint8_t* src, uint32_t bytes, bool isNumber);
-	void displayNumber(const uint8_t* src, uint32_t bytes);
-	void displayText(const uint8_t* src, uint32_t bytes);
+	void display(const uint32_t* src, uint32_t bytes);
+	void display(const uint32_t* src, uint32_t bytes, bool isNumber);
+	void displayNumber(const uint32_t* src, uint32_t bytes);
+	void displayText(const uint32_t* src, uint32_t bytes);
 };
 
 #endif
